Added a template to allow the HeapBlock class to be given signed ints or other types that are not size_t for its size parameters

+++ modules/juce_audio_basics/buffers/juce_FloatVectorOperations.cpp
@@ -1139,8 +1139,8 @@ public:
             const int range = random.nextBool() ? 500 : 10;
             const int num = random.nextInt (range) + 1;
 
-            HeapBlock<ValueType> buffer1 ((size_t) num + 16), buffer2 ((size_t) num + 16);
-            HeapBlock<int> buffer3 ((size_t) num + 16);
+            HeapBlock<ValueType> buffer1 (num + 16), buffer2 (num + 16);
+            HeapBlock<int> buffer3 (num + 16);
 
            #if JUCE_ARM
             ValueType* const data1 = buffer1;
